修改了 申请好友等几个bug

+++ app/controllers/add_group_applications_controller.rb
@@ -1,18 +1,14 @@
 class AddGroupApplicationsController < ApplicationController
+  before_filter :check_login?
+  before_filter :find_group_and_author_check, :except => [:apply]
   #TODO 添加 批量操作 能批量加入或批量忽略
   def index
     @page_title ="加入小组申请列表"
-    @group = Group.find_by_id(params[:group_id])
-    if @group
-      @add_applications = @group.add_applications.paginate :page => params[:page]
-    else
-      redirect_to groups_path()
-    end
+    @add_applications = @group.add_applications.paginate :page => params[:page]
   end
   #申请加入小组
   def apply
     @add_application = AddGroupApplication.new(params[:add_group_application])
-    
     if request.post?
       group = Group.find(params[:group_id])
       @add_application.applicant = current_user.get_account(params[:alias])
@@ -26,34 +22,32 @@ class AddGroupApplicationsController < ApplicationController
   end
 
   def accept
-    @group = Group.find_by_id(params[:group_id])
-    if @group
-      add_app = @group.add_applications.find_by_id(params[:id])
-      if add_app
-        if !@group.is_member?(add_app.applicant)
-          (@group.all_members << add_app.applicant)
-        end
-        add_app.destroy
-        flash[:success] = "接受了申请！"
-      else
-        flash[:error] = "产生错误！"
-      end
-      redirect_to :action=>"index"  , :page => params[:page]
+    add_app = @group.add_applications.find_by_id(params[:id])
+    if add_app && @group.add_to_member(add_app.applicant)
+      add_app.destroy
+      flash[:success] = "接受了申请！"
     else
-      redirect_to groups_path()
+      flash[:error] = "产生错误！"
     end
+    redirect_to :action=>"index"  , :page => params[:page]
+
   end
 
   def destroy
-    @group = Group.find_by_id(params[:group_id])
-    if @group
-      if  ! (@group.add_applications.find_by_id(params[:id]) && @group.add_applications.find_by_id(params[:id]).destroy)
-        flash[:error] = "产生错误！"
-      end
-      redirect_to :action=>"index"  , :page => params[:page]
-    else
-      redirect_to groups_path()
+    if  ! (@group.add_applications.find_by_id(params[:id]) && @group.add_applications.find_by_id(params[:id]).destroy)
+      flash[:error] = "产生错误！"
     end
+    redirect_to :action=>"index"  , :page => params[:page]
   end
+  
+  protected
+  #查找小组 并检查授权设置
+  def find_group_and_author_check
+    @group = Group.find(params[:group_id])
+    if not @group.is_manager_member?(current_user)
+      flash[:error] = @group.errors.full_messages.to_s
+      redirect_to @group
+    end
 
+  end
 end
